backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blockbuster==1.5.26
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
croniter==6.0.0
cryptography==46.0.4
distro==1.9.0
fastapi==0.128.2
filetype==1.2.0
forbiddenfruit==0.1.4
google-auth==2.48.0
google-genai==1.62.0
googleapis-common-protos==1.72.0
grandalf==0.8
grpcio==1.78.0
grpcio-health-checking==1.78.0
grpcio-tools==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain-core==1.2.9
langchain-google-genai==4.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-api==0.7.26
langgraph-checkpoint==4.0.0
langgraph-cli==0.4.12
langgraph-prebuilt==1.0.7
langgraph-runtime-inmem==0.24.0
langgraph-sdk==0.3.4
langsmith==0.6.9
neo4j==6.1.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.50.0
structlog==25.5.0
tavily-python==0.7.21
tenacity==9.1.3
tiktoken==0.12.0
truststore==0.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
